Update semantic scholar fetching mechanism to include exact publication date when available (#142)
* Update scripts to use publicationDate instead of year

* Remove old test data

* Update tested years

* Update tests to generate files with years

* Fix year fetching mechanism

* fix int casting

* Add newly generated files

* Update flag in test file

+++ tests/test_add_publications_by_author.py
@@ -6,6 +6,7 @@ from textwrap import dedent
 
 import src.python.add_publications_by_author as mod
 
+REMOVE_GENERATED_FILES = True
 
 class TestAddPublicationsByAuthor(unittest.TestCase):
     @classmethod
@@ -19,7 +20,7 @@ class TestAddPublicationsByAuthor(unittest.TestCase):
 
 
     def tearDown(self) -> None:
-        if os.path.exists(self.save_dir):
+        if REMOVE_GENERATED_FILES and os.path.exists(self.save_dir):
             shutil.rmtree(self.save_dir)
 
     def add_publication_and_verify_all(self, author):
